Initial Commit - working as intended
This is mostly just the SDL basic tutorial #1, modified to draw the
contents of a text file to the window. The only argument is the filename
of a file on disk to mmap, which will be read as though it were the
framebuffer in the S4548 kernel driver.
Other applications can then write to this file, and the resulting output
will be shown in the window.